Removed leftovers from testing
Sorry - my fault. 
I kept GITHUB_BRANCH as testing because otherwise this testing version will install not the version from the testing branch but the version from master which is inconsistent. But this requires to change the branch name when merging this branch into master. But I don't mind if you remove this variable (I granted you edit rights to my fork when I created the pull request ;-) )